fix(leads): the breach scan skips a lead nobody asked for, not one nobody owns
The scan narrowed to `owner_id IS NOT NULL`, which retired the configured
intake seat: an unowned INBOUND lead is the funnel's normal arrival state and
is exactly the row that seat answers for, so dropping every unowned row left
`unassigned_escalation_user_id` wired to nothing and a web-form lead nobody
picked up escalating to nobody, forever. Two tests already said so and went
red on main.

The concept the narrowing was reaching for is SOURCE, which is what the file
it came from says: a name the product read off a public web page is a record
that the person exists — nobody wrote in, so no first response is owed. The
sweep now asks that question, binding the kernel's own set rather than
spelling a second one, so a source the lead.created automations learn to
decline is one the sweep stops escalating in the same change.

Taking a lead on clears the breach stamp. That was the other half of the
argument for the ownership narrowing and it is a real defect on its own: the
scan admits only `sla_breached_at IS NULL`, so a lead escalated while it was
nobody's could never breach the NEW deadline that its pickup starts — the rep
who took it on could go quiet for a month with nobody told. Cleared under the
existing `routed_at IS NULL` guard, so it only ever clears a breach earned
while the row belonged to nobody.

The siteread fixture now writes source_system as the shipped accept path does;
it set only `source`, so it proved nothing about the column the rule reads.
